+++ src/utils/file/is-test-file.js
@@ -1,0 +1,6 @@
+// @flow
+
+export default function isTestFile(fileName: string) {
+  // see unit tests for examples
+  return new RegExp('^.*.(test|spec|specs|tests).(js|ts|tsx|jsx)$').test(fileName);
+}